Gradually improve GPS estimated H/V position accuracy after startup, set fix type based on current accuracy; Start/restart GPS at HILSystem Mavlink init instead of at program startup; Fix automatic restart issue when stopping simulator with (Q) key.

+++ src/me/drton/jmavsim/GNSSReport.java
@@ -9,8 +9,8 @@ import javax.vecmath.Vector3d;
  */
 public class GNSSReport {
     public LatLonAlt position;
-    public double eph;
-    public double epv;
+    public float eph;
+    public float epv;
     public Vector3d velocity;
     public int fix;     // 0 = no fix, 1 = time only, 2 = 2D fix, 3 = 3D fix
     public long time;   // UTC time in [us]
